Cache report

Print a report at the console with cache statistics

Usage

cache_report(cache_type = NULL, cache_dir = NULL)

Arguments

cache_type

character. Use temporary to a session temporary folder, persistent for a persistent cache folder or NULL to use the environment default. Check the section Cache Type for more details.

cache_dir

character. User specified cache directory for persistent cache type.

Value

No return value. The function prints a report on the console.

Examples

mirror_deposit(deposit_id = 10959197)
#> /tmp/Rtmpy6qK26/zendown/10959197
cache_report()
#> 
#> ── Cache report ────────────────────────────────────────────────────────────────
#> • Path: /tmp/Rtmpy6qK26/zendown
#> • Cached deposits: 1
#> • Total size: 3.94K
#> • Files: 5
#> 
#> ── Cached deposits ─────────────────────────────────────────────────────────────
#> 
#> ── Zenodo id: 10959197 ──
#> 
#> • Size: 3.94K
#> • Files: 5
#> • Creation time: 2024-10-17 16:02:43
#> • Modification time: 2024-10-17 16:02:44
#> • Access time: 2024-10-17 16:02:44