Get a list of files available at a Zenodo deposit

Get a list of files available at a Zenodo deposit

Usage

list_deposit(deposit_id)

Arguments

deposit_id

numeric. The Zenodo deposit id.

Value

a tibble.

Examples

# https://zenodo.org/records/10959197
list_deposit(deposit_id = 10959197)
#> # A tibble: 5 × 6
#>   id                                   filename filesize checksum self  download
#>   <chr>                                <chr>       <dbl> <chr>    <chr> <chr>   
#> 1 e09ba3f1-f9b5-44ad-a403-d084f046c96d mtcars.…     1225 994977c… http… https:/…
#> 2 1eee5ca9-272c-48e3-9962-ae4c72c1f1d9 ToothGr…      371 365f3f8… http… https:/…
#> 3 2873000e-40bc-44f5-ae0f-f6fcdcaff3a5 PlantGr…      325 d185729… http… https:/…
#> 4 b751722e-a6a8-47c5-b9b0-6d37afdab932 iris3.r…     1025 7006023… http… https:/…
#> 5 21ce7a06-effe-4c40-b8b7-4ed8da3075f2 iris.rds     1091 85dda31… http… https:/…